insurance-extraction-engine/services/paddleocr/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
paddlepaddle==2.5.2
paddleocr==2.7.0.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
loguru==0.7.2